scoring algorithm fix
Before this commit the scoring algorithm did not take into consideration that post-HSV masking **could** result in some grayscale tile color values being 0 and thus when subtracting 1 would wrap around and default to 255, trying to access the counts of property ids with index 255. This was corrected to use counts[property.id] instead and fixed it for good. Tiles that are not thresholded (usually the table top) will be ignored from the scoring process